Release 2024-04-11

+++ src/pages/api/fietsenstallingen_services/deleteForParking.ts
@@ -1,8 +1,8 @@
-import { Prisma } from "@prisma/client";
 import { prisma } from "~/server/db";
+import type { NextApiRequest, NextApiResponse } from "next";
 
-export default async function handle(req, res) {
-  if(! req.query.fietsenstallingId) return;
+export default async function handle(req: NextApiRequest, res: NextApiResponse) {
+  if (!req.query.fietsenstallingId || Array.isArray(req.query.fietsenstallingId)) return;
 
   const result = await prisma.fietsenstallingen_services.deleteMany({
     where: {
